requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
loguru==0.7.3
python-dotenv==1.0.1
pydantic==2.10.3
aiohttp==3.11.10
pydub==0.25.1
httpx==0.27.0
requests==2.31.0

